+++ src/main/java/online/kingdomkeys/kingdomkeys/command/ModCommands.java
@@ -18,7 +18,7 @@ public class ModCommands {
                 .then(FocusPointsCommand.register())
                 .then(HeartsCommand.register())
                 .then(LevelCommand.register())
-                .then(MagicLevelCommand.register())
+                //.then(MagicLevelCommand.register())
                 .then(MaterialCommand.register())
                 .then(MunnyCommand.register())
                 .then(PayMunnyCommand.register())